COMMERCE-1025 commerce-openapi-util - Reorganize - move property factory to ConfigurationFatory package
